style: add main styles
