+++ src/components/AnswerItems/AnswerItems.scss
@@ -2,6 +2,11 @@
   flex: auto;
   max-width: 47%;
 
+  @media (max-width: 768px) {
+    max-width: 100%;
+    width: 100%;
+  }
+
   &__list {
     border: 1px solid #000;
     border-radius: 5px;
