+++ src/components/App/App.js
@@ -2,10 +2,7 @@ import React from 'react';
 
 // components
 import Header from '../Header/Header';
-import Question from '../Question/Question';
-import AnswerItems from '../AnswerItems/AnswerItems';
-import Description from '../Description/Description';
-import NextButton from '../NextButton/NextButton';
+import Game from '../Game/Game';
 import Finish from '../Finish/Finish';
 
 // services
@@ -157,27 +154,6 @@ class App extends React.Component {
       </div>
     )
   }
-}
-
-const Game = ({data, onSelectItem, onClickNext}) => {
-  const { group, questionItem, selectedItem, guessed } = data;
-
-  return (
-    <main>
-      <Question
-        questionItem={questionItem}
-        guessed={guessed} />
-      <div className="items-block">
-        <AnswerItems
-          group={group}
-          onSelectItem={onSelectItem} />
-        <Description selectedItem={selectedItem} />
-      </div>
-      <NextButton
-        guessed={guessed}
-        onClickNext={onClickNext} />
-    </main>
-  )
 }
 
 export default App;
