+++ src/components/App/App.scss
@@ -4,20 +4,23 @@
   box-sizing: border-box;
 }
 
+body {
+  font-family: 'Roboto', sans-serif;
+  font-weight: 400;
+  color: #2d303a;
+  font-size: 16px;
+  background-color: khaki;
+}
+
 .app-container {
   max-width: 1300px;
   width: 100%;
   margin: 0 auto;
-  padding: 10px 20px;
-}
-
-.items-block {
-  display: flex;
-  justify-content: space-between;
-  flex: 0 0 50%;
-  margin-bottom: 30px;
+  padding: 15px 20px;
 
   @media (max-width: 768px) {
-    flex-direction: column;
+    max-width: 580px;
+    width: 100%;
+    padding: 15px 10px;
   }
 }
