+++ src/components/Description/Description.scss
@@ -10,6 +10,12 @@
   &__text {
     text-align: left;
   }
+
+  @media (max-width: 768px) {
+    max-width: 100%;
+    width: 100%;
+    margin-top: 30px;
+  }
 }
 
 .description-top {
@@ -18,11 +24,6 @@
 
   width: 100%;
   margin-bottom: 15px;
-
-  @media (max-width:1300px) {
-    display: block;
-    // max-width: 300px;
-  }
 
   &__img {
     width: 200px;
@@ -37,13 +38,20 @@
 
   &__name {
     font-size: 25px;
-    height: 35px;
+    padding-bottom: 10px;
     margin-bottom: 15px;
     border-bottom: 1px solid #000;
 
     &_en {
-      font-size: 16px;
-      height: 26px;
+      font-size: inherit;
+    }
+  }
+
+  @media (max-width:1300px) {
+    display: block;
+
+    &__info {
+      margin-top: 10px;
     }
   }
 }
