+++ src/components/Game/Game.scss
@@ -1,0 +1,11 @@
+.items-block {
+  display: flex;
+  justify-content: space-between;
+  flex: 0 0 50%;
+  margin-bottom: 30px;
+
+  @media (max-width: 768px) {
+    flex-direction: column;
+    align-items: center;
+  }
+}
