+++ src/components/Player/Player.scss
@@ -1,8 +1,32 @@
 .rhap_container {
   background-color: inherit;
   box-shadow: none;
+  padding: 0;
+  padding-bottom: 10px;
+}
+
+.rhap_controls-section {
+  flex: none;
+  margin-right: 10px !important;
 }
 
 .rhap_volume-controls {
-  min-width: 65px;
+  width: 70px;
+}
+
+.rhap_progress-section {
+  position: relative;
+}
+
+.rhap_time {
+  position: absolute;
+  top: 33px;
+}
+
+.rhap_total-time {
+  right: 5px;
+}
+
+.rhap_current-time {
+  left: 5px;
 }
