+++ src/components/Question/Question.scss
@@ -3,7 +3,7 @@
   justify-content: left;
 
   width: 100%;
-  padding: 15px 40px 15px 15px;
+  padding: 15px 80px 15px 15px;
   background-color: royalblue;
   border: 1px solid #000;
   border-radius: 5px;
@@ -12,11 +12,11 @@
   &__img {
     width: 200px;
     border-radius: 5px;
+    margin-right: 30px;
   }
 
   &__info{
     width: 100%;
-    margin-left: 30px;
   }
 
   &__name {
@@ -24,6 +24,18 @@
     padding-bottom: 25px;
     margin-bottom: 15px;
     border-bottom: 1px solid #000;
+  }
+
+  @media (max-width: 768px) {
+    padding: 15px;
+  }
+
+  @media (max-width: 600px) {
+    flex-direction: column;
+
+    &__info {
+      margin-top: 25px;
+    }
   }
 }
